+++ plugin_tooling/src-tests/com/googlecode/goclipse/tooling/env/GoEnvironmentTest.java
@@ -31,22 +31,22 @@ public class GoEnvironmentTest extends CommonGoToolingTest {
 	public void test_GoPath$() throws Exception {
 		GoPath goPath = new GoPath(WS_FOO + File.pathSeparator + WS_BAR);
 		 
-		assertAreEqual(goPath.getGoWorkspacePathEntry(WS_FOO.resolve("xxx")), WS_FOO);
-		assertAreEqual(goPath.getGoWorkspacePathEntry(WS_BAR.resolve("xxx")), WS_BAR);
-		assertAreEqual(goPath.getGoWorkspacePathEntry(TESTS_WORKDIR.resolve("xxx")), null);
+		assertAreEqual(goPath.findGoPathEntry(WS_FOO.resolve("xxx")), WS_FOO);
+		assertAreEqual(goPath.findGoPathEntry(WS_BAR.resolve("xxx")), WS_BAR);
+		assertAreEqual(goPath.findGoPathEntry(TESTS_WORKDIR.resolve("xxx")), null);
 		
-		assertAreEqual(goPath.getGoPackageFromSourceModule(WS_FOO.resolve("xxx/m.go")), null);
-		assertAreEqual(goPath.getGoPackageFromSourceModule(WS_FOO.resolve("src/xxx/m.go")), path("xxx"));
-		assertAreEqual(goPath.getGoPackageFromSourceModule(WS_FOO.resolve("src/xxx/zzz/m.go")), path("xxx/zzz"));
-		assertAreEqual(goPath.getGoPackageFromSourceModule(WS_FOO.resolve("src/m.go")), null);
-		assertAreEqual(goPath.getGoPackageFromSourceModule(WS_BAR.resolve("src/xxx/m.go")), path("xxx"));
-		assertAreEqual(goPath.getGoPackageFromSourceModule(WS_BAR.resolve("src/src/src/m.go")), path("src/src"));
-		assertAreEqual(goPath.getGoPackageFromSourceModule(TESTS_WORKDIR.resolve("src/xxx/m.go")), null);
+		assertAreEqual(goPath.findGoPackageForSourceModule(WS_FOO.resolve("xxx/m.go")), null);
+		assertAreEqual(goPath.findGoPackageForSourceModule(WS_FOO.resolve("src/xxx/m.go")), path("xxx"));
+		assertAreEqual(goPath.findGoPackageForSourceModule(WS_FOO.resolve("src/xxx/zzz/m.go")), path("xxx/zzz"));
+		assertAreEqual(goPath.findGoPackageForSourceModule(WS_FOO.resolve("src/m.go")), null);
+		assertAreEqual(goPath.findGoPackageForSourceModule(WS_BAR.resolve("src/xxx/m.go")), path("xxx"));
+		assertAreEqual(goPath.findGoPackageForSourceModule(WS_BAR.resolve("src/src/src/m.go")), path("src/src"));
+		assertAreEqual(goPath.findGoPackageForSourceModule(TESTS_WORKDIR.resolve("src/xxx/m.go")), null);
 		
 		// Test empty case
 		goPath = new GoPath("");
 		assertTrue(goPath.isEmpty());
-		assertTrue(goPath.getGoPathElements().size() == 0);
+		assertTrue(goPath.getGoPathEntries().size() == 0);
 		assertEquals(goPath.getGoPathWorkspaceString(), "");
 	}
 	
@@ -60,9 +60,9 @@ public class GoEnvironmentTest extends CommonGoToolingTest {
 		
 		Path goRoot_Source = goEnv.getGoRoot_Path().resolve("src/pkg");
 		
-		assertAreEqual(goEnv.getGoPackageFromSourceModule(goRoot_Source.resolve("pack/m.go")), path("pack"));
-		assertAreEqual(goEnv.getGoPackageFromSourceModule(goRoot_Source.resolve("pack/foo/m.go")), path("pack/foo"));
-		assertAreEqual(goEnv.getGoPackageFromSourceModule(goRoot_Source.resolve("../foo/m.go")), null);
+		assertAreEqual(goEnv.findGoPackageForSourceModule(goRoot_Source.resolve("pack/m.go")), path("pack"));
+		assertAreEqual(goEnv.findGoPackageForSourceModule(goRoot_Source.resolve("pack/foo/m.go")), path("pack/foo"));
+		assertAreEqual(goEnv.findGoPackageForSourceModule(goRoot_Source.resolve("../foo/m.go")), null);
 	}
 	
 }