+++ documentation/ChangeLog.md
@@ -1,6 +1,7 @@
 ## release ChangeLog
 
 ### Goclipse 0.8.x
+ * Fixed: Project Explorer no longer shows GOPATH entry if that entry location is the as the project location.
  * Added gocode process argument information to gocode log console.
  * Added Ctrl-click open definition functionality using `go oracle`.
    * Path to `oracle` command can be configured in the Go preference page.